Refactor process-document-headlessly to prevent circular deps
