+++ embeddable-say-editor/app/plugins/setup/setup-plugins.ts
@@ -120,12 +120,11 @@ export function setupPlugins(args: PluginInitArgs): EditorSetup {
     prosePlugins,
     schema,
     nodeViews: (controller: SayController) => {
-      const result: Record<string, NodeViewConstructor> = {};
+      const nvConstructors: Record<string, NodeViewConstructor> = {};
       for (const [name, nodeView] of Object.entries(nodeViews)) {
-        // TODO: these core types are weird, should be fixed in the editor
-        result[name] = nodeView(controller) as unknown as NodeViewConstructor;
+        nvConstructors[name] = nodeView(controller);
       }
-      return result;
+      return nvConstructors;
     },
     sidebarConfig: sidebar ?? defaultSidebar(args),
     toolbarConfig: toolbar ?? defaultToolbar(args),
